requirements.txt
docker>=6.0.0
fastapi==0.115.12
pydantic==2.11.3
PyYAML==6.0.2
uvicorn
pymongo